+++ apps/shared/avifjpeg.c
@@ -347,6 +347,8 @@ avifBool avifJPEGRead(const char * inputFilename,
                     goto cleanup;
                 }
                 avifImageSetMetadataExif(avif, marker->data + tagExif.size, marker->data_length - tagExif.size);
+                // Ignore any Exif parsing failure.
+                (void)avifImageExtractExifOrientationToIrotImir(avif);
                 found = AVIF_TRUE;
             }
         }